+++ packages/testing-helpers/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.22](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@0.7.21...@open-wc/testing-helpers@0.7.22) (2019-03-08)
+
+**Note:** Version bump only for package @open-wc/testing-helpers
+
+
+
+
+
 ## [0.7.21](https://github.com/open-wc/open-wc/compare/@open-wc/testing-helpers@0.7.20...@open-wc/testing-helpers@0.7.21) (2019-03-06)
 
 **Note:** Version bump only for package @open-wc/testing-helpers