+++ packages/webpack-import-meta-loader/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.1.2...@open-wc/webpack-import-meta-loader@0.1.3) (2019-03-08)
+
+**Note:** Version bump only for package @open-wc/webpack-import-meta-loader
+
+
+
+
+
 ## [0.1.2](https://github.com/open-wc/open-wc/compare/@open-wc/webpack-import-meta-loader@0.1.1...@open-wc/webpack-import-meta-loader@0.1.2) (2019-02-16)